refactor: split logic
